requirements.txt
openai>=1.0.0
akshare>=1.10.0
pyyaml>=6.0
schedule>=1.2.0
tavily-python>=0.3.0
